requirements.txt
absl-py==0.14.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
cycler==0.10.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958435052/work
filelock==3.2.0
flatbuffers==1.12
ftfy==6.0.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
huggingface-hub==0.0.17
idna==3.2
importlib-metadata==4.8.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
nltk==3.6.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681920064/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opt-einsum==3.3.0
packaging==21.0
Pillow @ file:///tmp/build/80754af9/pillow_1625670622947/work
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pynvml==11.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.96
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow @ file:///home/zhenli/code/tensorflow-2.6.2-cp36-cp36m-manylinux2010_x86_64.whl
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tqdm==4.62.3
transformers==4.10.0
typing-extensions==3.7.4.3
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.6.0
